deleted catalog.js
the reason why this was here in the first place was because I was using
a label for the file input, as styling a file input directly is fucking
impossible on all browsers. the two things it reimplements (drag-n-drop
and showing the file name) are already implemented by default in the
browsers for file inputs.